# server/src/main/resources/open-api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFireStationManager API",
...
    "contact": {
      "name": "OpenFireStationManager",
      "url": "https://github.com/simonhauck/OpenFireStationManager"
    },
    "version": "v1"
  },
  "security": [ {
    "sessionCookie": [ ]
  } ],
...
  "paths": {
...
    "/api/admin/privacy-policy": {
      "get": {
        "tags": [ "Admin - Legal" ],
        "summary": "Describe the stored privacy policy document",
        "description": "Returns the file name, MIME type, size, and upload time of the stored privacy policy (`Datenschutzerklärung`) — but not the file itself. Download the contents from `GET /privacy-policy`, which is public.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "getPrivacyPolicyMetadata",
        "responses": {
          "200": {
            "description": "Details of the stored document.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyMetadata"
                }
              }
            }
          },
          "404": {
            "description": "No privacy policy document has been uploaded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [ "Admin - Legal" ],
        "summary": "Upload the privacy policy document",
        "description": "Stores a privacy policy file as `multipart/form-data`. At most one document exists at a time, so uploading always **replaces** the previous one.\n\nOnly `application/pdf`, `text/html`, and `text/plain` are accepted; anything else is rejected with `422`. Files are capped at 10 MB by the server's multipart limits, and an oversized upload fails before this endpoint is reached.\n\nThis is the one place in the API that answers `201 Created` rather than `200`.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "uploadPrivacyPolicy",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The document to store. Must be a PDF, HTML, or plain-text file of at most 10 MB."
                  }
                },
                "required": [ "file" ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Details of the newly stored document.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyMetadata"
                }
              }
            }
          },
          "422": {
            "description": "The file's MIME type is not one of the three accepted types. Nothing was stored and any previous document is untouched.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "The request failed validation. The `errors` array names each offending field.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [ "Admin - Legal" ],
        "summary": "Remove the stored privacy policy document",
        "description": "Deletes the document, after which the public download returns `404`. Succeeds even when nothing is stored, so it is safe to call blindly.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "deletePrivacyPolicy",
        "responses": {
          "204": {
            "description": "No document is stored now."
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/privacy-policy": {
      "get": {
        "tags": [ "Legal" ],
        "summary": "Download the published privacy policy document",
        "description": "Streams the privacy policy (`Datenschutzerklärung`) file itself. The response is the raw document — a PDF, HTML, or plain-text body, not JSON — served with the stored MIME type and `Content-Disposition: inline` so browsers display rather than download it.\n\nNote the path: this endpoint sits at the site root, **outside** the `/api` namespace, because it is linked to directly from the website footer.\n\n**Authentication:** none — this endpoint is public.",
        "operationId": "downloadPrivacyPolicy",
        "responses": {
          "200": {
            "description": "The document contents, in the MIME type it was uploaded with.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "No privacy policy document has been uploaded.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/public/privacy-policy/exists": {
      "get": {
        "tags": [ "Legal" ],
        "summary": "Check whether a privacy policy is published",
        "description": "Returns a boolean so a page can decide whether to show a privacy policy link before attempting the download.\n\n**Authentication:** none — this endpoint is public.",
        "operationId": "publicPrivacyPolicyExists",
        "responses": {
          "200": {
            "description": "Whether a document exists.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyExists"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
...
    "/api/admin/privacy-policy/exists": {
      "get": {
        "tags": [ "Admin - Legal" ],
        "summary": "Check whether a privacy policy document is stored",
        "description": "Returns a boolean without the document or its details. Use it to decide whether to call `getPrivacyPolicyMetadata`, which fails with `404` when nothing is stored.\n\n**Authorisation:** requires the `ADMIN` role. An `ADMIN` account implicitly holds every role and may also call this.",
        "operationId": "privacyPolicyExists",
        "responses": {
          "200": {
            "description": "Whether a document exists.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PrivacyPolicyExists"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated — no valid session cookie was supplied.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but the account lacks the required role.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error. The response body carries no details.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": { }
          }
        }
      },
      "ValidationError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "type": "string",
            "description": "Path of the offending field within the request body, e.g. `size`. Empty when the rule applies to the request as a whole rather than one field.",
            "example": "size"
          },
          "message": {
            "type": "string",
            "description": "Human-readable reason the value was rejected.",
            "example": "must not be blank"
          },
          "rejectedValue": {
            "type": "object",
            "description": "The value that was submitted and rejected. May be null."
          }
        }
      },
      "ValidationProblemDetail": {
        "description": "RFC 9457 problem document describing a request that failed validation, extended with an `errors` array naming each offending field.",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Invalid request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Validation failed for request body"
          },
          "instance": {
            "type": "string",
            "description": "URI of the request that failed, when available."
          },
          "errors": {
            "type": "array",
            "description": "Every field-level failure detected in this request.",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "PrivacyPolicyMetadata": {
        "type": "object",
        "description": "Details about the stored privacy policy file, without its contents. Download the file itself from `GET /privacy-policy`.",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "Original name of the uploaded file.",
            "example": "datenschutzerklaerung.pdf"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type of the stored file — one of `application/pdf`, `text/html`, or `text/plain`.",
            "example": "application/pdf"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "description": "Size of the file in bytes.",
            "example": 254118
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file was uploaded."
          }
        },
        "required": [ "contentType", "fileName", "fileSize", "uploadedAt" ]
      },
...
      "PrivacyPolicyExists": {
        "type": "object",
        "description": "Whether a privacy policy document has been uploaded. Lets a caller check without downloading the file.",
        "properties": {
          "exists": {
            "type": "boolean",
            "description": "True when a document is stored.",
            "example": true
          }
        },
        "required": [ "exists" ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Server-side session cookie issued by `POST /api/public/auth/login`. Send it on every authenticated request.",
        "name": "OFSM_AUTH",
        "in": "cookie"
      },
      "rememberMeCookie": {
        "type": "apiKey",
        "description": "Long-lived cookie issued when logging in with `rememberMe = true`, valid for 30 days. It authenticates on its own, so a long-lived client may send it instead of a session cookie.",
        "name": "OFSM_AUTH_REMEMBER_ME",
        "in": "cookie"
      }
    }
  }
}